Solution to Find Maximum of Three Numbers

+++ Find Maximum of Three Numbers@@6788dd63490637cfc44e1b94/code.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 int main(){
-    float a,b,c;
-    scanf("%f %f %f",&a,&b,&c);
+    int a,b,c;
+    scanf("%d %d %d",&a,&b,&c);
     if(a>b && a>c){
-        printf("%f",a);
+        printf("%d",a);
     }else if(b>a && b>c){
-        printf("%f",b);
+        printf("%d",b);
     }else if(c>a && c>b){
-        printf("%f",c);
+        printf("%d",c);
     }
     return 0;
 }
